+++ src/main/java/com/rapidtect/springrestapi/repository/CategoryRepo.java
@@ -4,5 +4,5 @@ import com.rapidtect.springrestapi.entity.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.support.JpaRepositoryImplementation;
 
-public interface CategoryRepo extends JpaRepository<CategoryEntity, Integer> {
+public interface CategoryRepo extends JpaRepository<CategoryEntity, Long> {
 }
