+++ src/main/java/com/rapidtect/springrestapi/repository/CustomerRepo.java
@@ -3,5 +3,5 @@ package com.rapidtect.springrestapi.repository;
 import com.rapidtect.springrestapi.entity.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepo extends JpaRepository<CustomerEntity, Integer> {
+public interface CustomerRepo extends JpaRepository<CustomerEntity, Long> {
 }
