+++ src/main/java/com/rapidtect/springrestapi/repository/PurchaseOrderDetailRepo.java
@@ -3,5 +3,5 @@ package com.rapidtect.springrestapi.repository;
 import com.rapidtect.springrestapi.entity.PurchaseOrderDetailEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseOrderDetailRepo extends JpaRepository<PurchaseOrderDetailEntity, Integer> {
+public interface PurchaseOrderDetailRepo extends JpaRepository<PurchaseOrderDetailEntity, Long> {
 }
