+++ src/main/java/com/rapidtect/springrestapi/service/CategoryService.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 
 public interface CategoryService {
     List<CategoryModel> getAll();
-    Optional<CategoryModel> getById(Integer id);
+    Optional<CategoryModel> getById(Long id);
     Optional<CategoryModel> save(CategoryModel model);
-    Optional<CategoryModel> update(Integer id, CategoryModel model);
-    Optional<CategoryModel> delete(Integer id);
+    Optional<CategoryModel> update(Long id, CategoryModel model);
+    Optional<CategoryModel> delete(Long id);
 }
