+++ src/main/java/com/rapidtect/springrestapi/service/CustomerService.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 
 public interface CustomerService {
     List<CustomerModel> getAll();
-    Optional<CustomerModel> getById(Integer id);
+    Optional<CustomerModel> getById(Long id);
     Optional<CustomerModel> save(CustomerModel model);
-    Optional<CustomerModel> update(Integer id, CustomerModel model);
-    Optional<CustomerModel> delete(Integer id);
+    Optional<CustomerModel> update(Long id, CustomerModel model);
+    Optional<CustomerModel> delete(Long id);
 
 }
